move get_current_context_template_data function to a proper place
